Add files via upload
This is a Database Project using HTML, CSS, Bootstrap,  MySQL and php. This project is actually based on an airplane management system that creates passengers or staffs account and give them options to add and check their details . The admin has access to all of the options like manage staff, manage flights etc. The database is connected at backend using php.